requirements.txt
accelerate==1.3.0
bert-score==0.3.13
datasets==3.2.0
evaluate==0.4.3
huggingface-hub==0.28.1
langchain==0.3.19
langdetect==1.0.9
langfuse==2.59.2
pandas==2.2.3
python-dotenv==1.0.1
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
